# Send logs concurrently