--- a/aws/logs_monitoring/tests/test_extract_metric.py
+++ b/aws/logs_monitoring/tests/test_extract_metric.py
@@ -8,6 +8,7 @@ sys.modules["datadog_lambda.wrapper"] = MagicMock()
 sys.modules["datadog_lambda.metric"] = MagicMock()
 sys.modules["datadog"] = MagicMock()
 sys.modules["requests"] = MagicMock()
+sys.modules["requests_futures.sessions"] = MagicMock()
 
 env_patch = patch.dict(
     os.environ,