--- a/aws/logs_monitoring/tests/test_invoke_additional_target_lambdas.py
+++ b/aws/logs_monitoring/tests/test_invoke_additional_target_lambdas.py
@@ -11,6 +11,7 @@ sys.modules["datadog_lambda.wrapper"] = MagicMock()
 sys.modules["datadog_lambda.metric"] = MagicMock()
 sys.modules["datadog"] = MagicMock()
 sys.modules["requests"] = MagicMock()
+sys.modules["requests_futures.sessions"] = MagicMock()
 
 env_patch = patch.dict(
     os.environ,