### What does this PR do?

Send AWS logs concurrently for better performance over handling larger log files (e.g., vpc log files from s3). Also increased the batch size to reduce # of batches.

CFN parameter `DdMaxWorkers` can be used to fine control the concurrency.

### Motivation

Forwarder may time out when handling large log files.

### Testing Guidelines

For a large vpc log file containing 800k log lines, duration reduced by more than 10X while memory usage remained the same

**Before:**
REPORT RequestId: XXX Duration: 650270.61 ms Billed Duration: 650300 ms Memory Size: 3008 MB Max Memory Used: 1454 MB

**After:**
REPORT RequestId: XXX Duration: 42998.54 ms	Billed Duration: 43000 ms Memory Size: 3008 MB Max Memory Used: 1503 MB

### Additional Notes

<!--- Anything else we should know when reviewing? --->

### Types of changes

- [ ] Bug fix
- [ ] New feature
- [ ] Breaking change
- [x] Misc (docs, refactoring, dependency upgrade, etc.)

### Check all that apply

- [x] This PR's description is comprehensive
- [ ] This PR contains breaking changes that are documented in the description
- [ ] This PR introduces new APIs or parameters that are documented and unlikely to change in the foreseeable future
- [ ] This PR impacts documentation, and it has been updated (or a ticket has been logged)
- [ ] This PR's changes are covered by the automated tests
- [ ] This PR collects user input/sensitive content into Datadog
- [ ] This PR passes the integration tests (ask a Datadog member to run the tests)
- [ ] This PR passes the unit tests 
- [x] This PR passes the installation tests (ask a Datadog member to run the tests)