@nhinsch 2020-11-24
I would add the new environment variable/parameter to the README, and I would also briefly explain why the customer might want to adjust the number of workers (I assume adding more workers means logs are forwarded faster, but the function uses more memory?).

Also, is it strictly necessary to expose the setting for the number of workers? If there is a value that you think would work for all customers, could we just hardcode that number of workers for greater simplicity?